log when train data is subsampled
There was a warning when subsampling option was choosen but data size
was not big enough to be subsampled. It makes more sense to log that not
all data was used. This happens now on the info level. It is not a
warning since the config option was actively chosen so no real reason to
warn about it.
The logic was changed such that the sampling only takes place when
train data size is strictly bigger than the sample size. Before it would
have been sampled on equality. This has the chance to chance some
results(when subset_size==data_size) since before the idx where sampled
and now the idx are sorted.