fix indeterminism
get_rows_from_dataframe is called in parallel and will call list-append
concurrently on the same list. This is thread safe but the ordering
depends on the thread speed. To fix this the original ordering has to
be restored. To do this the individual train indexes per file are
stacked so that the dimensions are the same as for the concatenated
chunked dfs. The concatenated df then can be resorted by the original
index.